+++ app/controllers/iterations_controller.rb
@@ -2,8 +2,8 @@ class IterationsController < ApplicationController
   def show
     authorize project
     iteration =
-      IterationOperations::Read.call(start_date: params.require(:start_date).to_time,
-                                     end_date: params.require(:end_date).to_time,
+      IterationOperations::Read.call(start_date: params.require(:start_date).to_date,
+                                     end_date: params.require(:end_date).to_date,
                                      project: project)
 
     render json: iteration