+++ app/javascript/packs/app.js
@@ -1,13 +1,16 @@
 import 'libs';
 import ProjectsIndexController from 'controllers/projects/IndexController';
+import ProjectsEditController from 'controllers/projects/EditController';
 import TagGroupsController from 'controllers/tag_groups/TagGroupsController';
+
+
 import { start as ProjectsShowController } from 'central';
 import BetaShowProjectBoardController from 'controllers/beta/project_boards/ShowController';
 
 const routes = {
   'projects.index': ProjectsIndexController,
   'projects.show': ProjectsShowController,
-  'projects.edit': TagGroupsController,
+  'projects.edit': ProjectsEditController,
   'tag_groups.new': TagGroupsController,
   'tag_groups.edit': TagGroupsController,
   'beta/projects.show': BetaShowProjectBoardController